+++ src/main/java/de/schmiereck/prednet/service/CurveDto.java
@@ -1,25 +1,25 @@
 package de.schmiereck.prednet.service;
 
 public class CurveDto {
-    private final int[] inputArr;
-    private final int[] outputHistorieArr;
-    private final int output;
+    private final long[] inputArr;
+    private final long[] outputHistorieArr;
+    private final long output;
 
-    public CurveDto(final int[] inputArr, final int[] outputHistorieArr, final int output) {
+    public CurveDto(final long[] inputArr, final long[] outputHistorieArr, final long output) {
         this.inputArr = inputArr;
         this.outputHistorieArr = outputHistorieArr;
         this.output = output;
     }
 
-    public int[] getInputArr() {
+    public long[] getInputArr() {
         return this.inputArr;
     }
 
-    public int[] getOutputHistorieArr() {
+    public long[] getOutputHistorieArr() {
         return this.outputHistorieArr;
     }
 
-    public int getOutput() {
+    public long getOutput() {
         return this.output;
     }
 }
